Help static analyzers find exported symbols
Define __all__ in a static manner to help static analysis tools. This is
specifically relevant to PyRight and its heuristic that determines which
symbols are exported by a module.

https://github.com/microsoft/pyright/blob/main/docs/typed-libraries.md

The doc string table of contents is now dynamically checked in a
different manner so that we still ensure that each member of __all__ is
represented in the ToC.