Repair sphinx documentation build
Some bitrot has caused the Sphinx-based documentation build on
readthedocs.io to fail for quite some time.

To repair, we add docs/requirements.txt which defines a contemporary
version of Sphinx along with its transient dependencies. The
.readthedocs.yaml file is updated to use the new docs/requirements.txt,
which overrides the default/old Sphinx [1].

Another type-checking circular dependency workaround is added to conf.py.

Also update to the new intersphinx_mapping format from the old format that
was deprecated in Sphinx 6.2.0.

And finally, add a new tox environment that may be used to perform the
sphinx documentation build in a sandboxed environment. The .gitlab-ci.yml
is updated to run this new tox build so that documentation regressions can
be caught in gitlab ci and not just rely on the readthedocs.io build.

[1]: https://docs.readthedocs.io/en/stable/build-default-versions.html