feat!: drop support for EOL Python 3.6 and 3.7
Python 3.6 was EOL'ed by the Python project as of 2021-12-23.
Python 3.7 was EOL as of 2023-06-27.

Dropping SimPy support for these EOL Python versions reduces the
maintenance burden, allows for changes that would be incompatible with
these older Python versions, and makes it easier to support newer Python
versions and packaging standards

SimPy users that may need support for Python 3.6 or 3.7 may continue to
use SimPy 4.0.2.