fix: possible unbound variable in _describe_frame
The `line` variable is bound in the for statement and thus could
conceivable never be bound (given an empty source file). Or the for loop
could fall through such that the last line of the file was used instead
of the target line number.

These are both highly theoretical possibilities, but the former triggers
pyright and it is easy enough to have more provably correct behavior.