fix: suppress some pyright errors
The .append and .remove operations annotated in this change confuse the
typechecker because the PutQueue and GetQueue ClassVar types do not
fully reconcile with the type annotations on the put_queue and get_queue
instance attributes.

PutQueue and GetQueue cannot be annotated perfectly due to a limitation
with ClassVar where it is disallowed from being used with a TypeVar.

I.e. this would be a more correct annotation:

    PutQueue: ClassVar[Type[MutableSequence[PutType]]]

But we can't write that, so the best we can do is:

    PutQueue: ClassVar[Type[MutableSequence]]