test: enable ruff pep8-naming checks
The SimPy code largely already follows pep-8 naming conventions. A few
names in a few tests are fixed-up for conformance and N818 is explicitly
disabled.

N818 mandates that names of exception types end with "Error". But SimPy
has several public exception types that are not named this way and to do
so would be an API-breaking change.